requirements.txt
diagrams
pyyaml